+++ CHANGELOG.md
@@ -84,6 +84,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - Fixed the bug that `/alwaysday` command actually do the opposite thing.
 - Several bugs in physics engine, including wrong collision detection and wrong movement calculation are fixed.
 - Gamerule `doDaylightCycle` now works correctly.
+- Fixed touch-mode player (phone player usually) block breaking with "delayed block breaking" option enabled
 
 ### Removed